+++ src/main/java/com/example/rublr/dummy/InMemoryRecordStore.java
@@ -22,7 +22,7 @@ public class InMemoryRecordStore implements RecordStore {
   }
 
   @Override
-  public int updateRecords(String name, List<BlogPost> delta) {
+  public long updateRecords(String name, List<BlogPost> delta) {
     Preconditions.checkNotNull(name);
     if (posts.putAll(name, delta)) {
       return delta.size();
